refactor lint code
