+++ src/momentOfTruthUtils.ts
@@ -3,8 +3,10 @@
 
 import * as stringMap from '@ts-common/string-map'
 import * as path from 'path'
-import * as fs from 'fs'
 import * as yaml from "js-yaml";
+import * as format from "@azure/swagger-validation-common";
+import * as utils from "./utils";
+import * as fs from "fs-extra";
 
 export type Issue = {
   readonly type?: string
@@ -59,7 +61,7 @@ export class LintingResultParser {
   }
 
   getResult() {
-    let results: any[] = [];
+    const results: any[] = [];
     let matches;
     while ((matches = this.regexLintResult.exec(this.results))) {
       try {
@@ -107,7 +109,7 @@ export type FinalResult = {
 
 // Creates and returns path to the logging directory
 export function getLogDir() {
-  let logDir = path.resolve('output');
+  const logDir = path.resolve('output');
   if (!fs.existsSync(logDir)) {
       try {
           fs.mkdirSync(logDir);
@@ -116,4 +118,75 @@ export function getLogDir() {
       }
   }
   return logDir;
+}
+
+export type Mutable<T extends object> = {
+  -readonly [K in keyof T]: T[K];
+};
+
+export type MutableIssue = Mutable<Issue>;
+
+
+export function getLine(jsonRef: string): number | undefined {
+  try {
+    return parseInt(
+      jsonRef.substr(jsonRef.indexOf(".json:") + 6).split(":")[0]
+    );
+  } catch (error) {
+    return undefined;
+  }
+}
+
+export function getFile(jsonRef: string) {
+  try {
+    const start = jsonRef.indexOf("specification");
+    return jsonRef.substr(start, jsonRef.indexOf(".json") + 5 - start);
+  } catch (error) {
+    return undefined;
+  }
+}
+
+export function getDocUrl(id: string) {
+  return `https://github.com/Azure/azure-rest-api-specs/blob/master/documentation/openapi-authoring-automated-guidelines.md#${id}`;
+}
+
+export function composeLintResult(it: MutableIssue) {
+  const severityMap: Map<string, string> = new Map([
+    ["error", "Error"],
+    ["warning", "Warning"],
+    ["info", "Info"],
+  ]);
+
+  const type = severityMap.get(String(it.type).toLowerCase())
+    ? severityMap.get(String(it.type).toLowerCase())
+    : "Info";
+  return {
+    level: type as format.MessageLevel,
+    message: String(it.message).replace(/"/g, "'"),
+    code: String(it.code),
+    id: String(it.id),
+    docUrl: getDocUrl(it.id),
+    time: new Date(),
+    extra: {
+      validationCategory: it.validationCategory,
+      providerNamespace: it.providerNamespace,
+      resourceType: it.resourceType,
+      jsonref: it.jsonref,
+      filePath: it.filePath,
+      lineNumber: it.lineNumber,
+      sources: it.sources,
+    },
+    paths: [
+      {
+        tag: "New",
+        path: utils.blobHref(
+          utils.getGithubStyleFilePath(
+            utils.getRelativeSwaggerPathToRepo(
+              it.filePath + "#L" + String(it.lineNumber) || ""
+            )
+          )
+        ),
+      },
+    ],
+  };
 }
